refactor(dockview-core): move module contracts into core (separate from impl)
Relocate the service + host interfaces for ContextMenu, TabGroupChips and
Accessibility out of their implementation files into a single core
`moduleContracts.ts`. Core (the ServiceCollection slots in modules.ts and the
host implementation in dockviewComponent.ts) now references the contracts from
core, never from the implementation files. Each implementation imports its own
contract from core.

Pure type relocation — no behaviour change, all 1117 tests green. This makes
the module boundary explicit: the contract is core, the implementation is a
detail that can be relocated without core depending on it.
